requirements.txt

streamlit
langchain
langgraph
langchain_openai
python-dotenv
wikipedia
